requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile ./requirements.in
#
attrs==19.1.0
backoff==1.8.0
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0
datadog==0.29.3
ddt==1.2.1                # via gitpython
decorator==4.4.0          # via datadog
gitdb2==2.0.5             # via gitpython
gitdb==0.6.4              # via gitpython
gitpython==2.1.12
idna==2.8                 # via requests
pybreaker==0.5.0
requests==2.22.0
six==1.12.0               # via pybreaker
smmap2==2.0.5             # via gitdb2
smmap==0.9.0              # via gitdb
urllib3==1.25.3           # via requests
